:recycle: [Refactor] ERD 수정 내용 적용
